+++ src/main/java/com/example/eCommerce/repository/ProductRepository.java
@@ -3,11 +3,10 @@ package com.example.eCommerce.repository;
 import com.example.eCommerce.model.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductRepository extends MongoRepository<Product, String> {
     List<Product> findByNameContainingIgnoreCase(String name);
     List<Product> findByCategoryIgnoreCase(String category);
-    List<Product> findByPriceBetween(BigDecimal min, BigDecimal max);
+    List<Product> findByPriceBetween(double min, double max);
 }
