+++ src/main/java/com/example/eCommerce/service/ProductService.java
@@ -2,7 +2,6 @@ package com.example.eCommerce.service;
 
 import com.example.eCommerce.model.Product;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
@@ -13,5 +12,5 @@ public interface ProductService {
     List<Product> getAllProducts();
     List<Product> searchByName(String name);
     List<Product> filterByCategory(String category);
-    List<Product> filterByPriceRange(BigDecimal minPrice, BigDecimal maxPrice);
+    List<Product> filterByPriceRange(double minPrice, double maxPrice);
 }
